Update P25Reflector.ini Name, FileLevel, FilePath
Revisions to default ini file to reduce the need for the user to edit the /opt/P25Reflector.ini 
Port may need to be changed IF you do not intend to use the default UDP41000.  Remember to open UDP41000 (or the port you choose) to the Intranet IP of the PC running this P25Reflector.